// src/api/generated/model/listUserAudiencesApiV1AudiencesMyGetParams.ts
/**
 * Generated by orval v8.8.0 🍺
 * Do not edit manually.
 * Amrita ICPC Coding Platform - Backend
 * A backend tool for Amrita ICPC Coding Platform
 * OpenAPI spec version: 1.0.0
 */

export type ListUserAudiencesApiV1AudiencesMyGetParams = {
    /**
     * Page number (starts from 1)
     * @minimum 1
     */
    page?: number;
    /**
     * Number of items per page
     * @minimum 1
     * @maximum 100
     */
    page_size?: number;
    /**
     * Optional search by audience name
     */
    q?: string | null;
};